Convert core module to package
