+++ divik/core/_utils.py
@@ -1,0 +1,37 @@
+from contextlib import contextmanager
+# RawArray exists, but PyCharm goes crazy
+# noinspection PyUnresolvedReferences
+from multiprocessing import Pool, RawArray
+
+import numpy as np
+from skimage.color import label2rgb
+
+from ._types import Data
+
+
+def normalize_rows(data: Data) -> Data:
+    normalized = data - data.mean(axis=1)[:, np.newaxis]
+    norms = np.sum(np.abs(normalized) ** 2, axis=-1, keepdims=True)**(1./2)
+    normalized /= norms
+    return normalized
+
+
+def visualize(label, xy, shape=None):
+    x, y = xy.T
+    if shape is None:
+        shape = np.max(y) + 1, np.max(x) + 1
+    y = y.max() - y
+    label = label - label.min() + 1
+    label_map = np.zeros(shape, dtype=int)
+    label_map[y, x] = label
+    image = label2rgb(label_map, bg_label=0)
+    return image
+
+
+@contextmanager
+def context_if(condition, context, *args, **kwargs):
+    if condition:
+        with context(*args, **kwargs) as c:
+            yield c
+    else:
+        yield None
